Move extra parameters indicator type check
Add a utility module for checking which set of extra query parameters, if any, a given indicator uses.
