+++ src/app/sidebar/sidebar.component.ts
@@ -3,6 +3,8 @@ import { Component, EventEmitter, ViewEncapsulation, OnInit, Output } from '@ang
 import { IndicatorService } from '../services/indicator.service';
 import { Indicator } from '../models/indicator.model';
 
+import { isThresholdIndicator } from '../charts/extra-params-components/extra-params.constants';
+
 /*
  * Sidebar Component
  * Populates sidebar with indicators and triggers adding charts
@@ -36,7 +38,7 @@ export class SidebarComponent implements OnInit {
 
         // TODO: Remove this filter once we've implemented user controls for all indicator parameters
         const visibleIndicators = indicators.filter(i => {
-            return i.thresholdIndicator || !i.parameters.some(p => p.required);
+            return isThresholdIndicator(i.name) || !i.parameters.some(p => p.required);
         });
 
         this.tempIndicators = visibleIndicators.filter(i => {
